+++ contracts/multisig-improved/src/action_types/discard.rs
@@ -9,37 +9,16 @@ pub trait DiscardActionModule:
     + super::sign::SignModule
     + super::perform::PerformModule
     + super::execute_action::ExecuteActionModule
-    + super::callbacks::CallbacksModule
+    + crate::ms_endpoints::callbacks::CallbacksModule
     + crate::check_signature::CheckSignatureModule
     + crate::external::events::EventsModule
 {
-    /// Clears storage pertaining to an action that is no longer supposed to be executed.
-    /// Any signatures that the action received must first be removed, via `unsign`.
-    /// Otherwise this endpoint would be prone to abuse.
-    #[endpoint(discardAction)]
-    fn discard_action_endpoint(&self, action_id: ActionId) {
-        let (_, caller_role) = self.get_caller_id_and_role();
-        caller_role.require_can_discard_action::<Self::Api>();
-
-        self.discard_action(action_id);
-    }
-
-    /// Discard all the actions with the given IDs
-    #[endpoint(discardBatch)]
-    fn discard_batch(&self, action_ids: MultiValueEncoded<ActionId>) {
-        let (_, caller_role) = self.get_caller_id_and_role();
-        caller_role.require_can_discard_action::<Self::Api>();
-
-        for action_id in action_ids {
-            self.discard_action(action_id);
-        }
-    }
-
     fn discard_action(&self, action_id: ActionId) {
         require!(
             self.get_action_valid_signer_count(action_id) == 0,
             "cannot discard action with valid signatures"
         );
+        
         self.abort_batch_of_action(action_id);
         self.clear_action(action_id);
     }
