+++ contracts/multisig-improved/src/action_types/mod.rs
@@ -1,8 +1,5 @@
-pub mod callbacks;
 pub mod discard;
 pub mod execute_action;
 pub mod perform;
-pub mod perform_endpoints;
 pub mod propose;
-pub mod propose_endpoints;
 pub mod sign;
