+++ contracts/multisig-improved/src/ms_endpoints/mod.rs
@@ -1,0 +1,5 @@
+pub mod callbacks;
+pub mod discard_endpoints;
+pub mod perform_endpoints;
+pub mod propose_endpoints;
+pub mod sign_endpoints;
